Add 11 embedded headers (stdlib/string/limits/errno/fcntl/signal/unistd/time/dirent/sys/*); fix int literal overflow and safeneg truncation
- Add stdlib.h, string.h, limits.h, errno.h, fcntl.h, signal.h,
  dirent.h, unistd.h, time.h, sys/wait.h, sys/stat.h as embedded
  headers. Constants via #define, functions via dlsym fallback.
  unistd.h and time.h provide type aliases (pid_t, size_t, time_t, etc.)
  via #define substitution.
- Update gen_embedded_headers.sh to use find+relative paths so
  sys/wait.h and sys/stat.h key correctly (was basename-only).
- Fix: decimal integer literals were stored in int (32-bit), causing
  overflow for values >= 2^31. Changed int v -> int64_t v in lexer and
  int _token -> int64_t _token in TokenBase/TokenInt.
- Fix: safeneg() had a spurious movsx(op, op.r8()) after cc.neg() that
  sign-extended from 8 bits, corrupting any negated value >= 128 in
  magnitude.

+++ include/datatokens.h
@@ -166,11 +166,11 @@ public:
     Variable &var;
     TokenVar(Variable &v) : TokenBase(), var(v) { _datatype = v.type; }
     virtual TokenType type() const { return TokenType::ttVariable; }
-    virtual int get() const { return var.get<int>(); }
-    virtual int val() const { return var.get<int>(); }
+    virtual int64_t get() const { return var.get<int64_t>(); }
+    virtual int val() const     { return var.get<int>(); }
     virtual bool is_constant() { return var.is_constant(); }
     virtual bool is_real() { return _datatype->is_real(); }
-    virtual void set(int c) { DBG(std::cout << "TokenVariable: set() calling var.set()" << std::endl); var.set(c); }
+    virtual void set(int64_t c) { DBG(std::cout << "TokenVariable: set() calling var.set()" << std::endl); var.set(c); }
     virtual void putreg(Program &);
 //  virtual asmjit::x86::Gp &getreg(Program &);
     virtual asmjit::Operand &operand(Program &);
